AG-14341 Axis formatter domain
